bigquery-proxy/requirements.txt
flask
flask-cors
flask-limiter
google-cloud-bigquery
google-cloud-storage
functions-framework
